requirements.txt
torch
torchtext
torchvision

flask
img2pdf
ocrmypdf
pillow
flask-ngrok
tabula-py
layoutparser
opencv-python
pytesseract
layoutparser[ocr]
tesseract
